added recursive fetchgin of subfolders and deletion of folder and its subfolders

+++ server/src/services/FolderServices.ts
@@ -7,14 +7,28 @@ export async function getAllFoldersService() {
   return folders;
 }
 export async function getSingleFolderService(folderId: string) {
-  const folder = await client.folder.findUnique({
-    where: {
-      id: folderId,
-    },
-    include: {
-      subFolders: true,
-    },
-  });
+  async function getFolderWithSubfolder(folderId: string) {
+    const folder = await client.folder.findUnique({
+      where: {
+        id: folderId,
+      },
+      include: {
+        subFolders: true,
+      },
+    });
+
+    if (!folder) return null;
+
+    folder.subFolders = await Promise.all(
+      folder.subFolders.map(async (folder) => {
+        return (await getFolderWithSubfolder(folder.id))!;
+      })
+    );
+    return folder;
+  }
+
+  const folder = getFolderWithSubfolder(folderId);
+
   return folder;
 }
 
@@ -29,10 +43,32 @@ export async function createFolderService({ name, parentFolderId }: { name: stri
 }
 
 export async function deleteFolderService(folderId: string) {
-  const deleteFolder = await client.folder.delete({
-    where: {
-      id: folderId,
-    },
-  });
-  return deleteFolder;
+  async function deleteFolderAndSubfolders(folderId: string) {
+    const folder = await client.folder.findUnique({
+      where: {
+        id: folderId,
+      },
+      include: {
+        subFolders: true,
+      },
+    });
+
+    if (!folder) return;
+
+    await Promise.all(
+      folder.subFolders.map((folder) => {
+        return deleteFolderAndSubfolders(folder.id);
+      })
+    );
+
+    await client.folder.delete({
+      where: {
+        id: folderId,
+      },
+    });
+  }
+
+  const deletedFolders = await deleteFolderAndSubfolders(folderId);
+
+  return deletedFolders;
 }
